+++ app/src/main/java/org/openedx/app/di/ScreenModule.kt
@@ -79,7 +79,7 @@ val screenModule = module {
         )
     }
     viewModel { (courseId: String?) ->
-        SignUpViewModel(get(), get(), get(), get(), get(), courseId)
+        SignUpViewModel(get(), get(), get(), get(), get(), get(), courseId)
     }
     viewModel { RestorePasswordViewModel(get(), get(), get(), get()) }
