+++ auth/src/main/java/org/openedx/auth/presentation/signup/SignUpUIState.kt
@@ -2,10 +2,12 @@ package org.openedx.auth.presentation.signup
 
 import org.openedx.core.domain.model.RegistrationField
 
-sealed class SignUpUIState {
-    data class Fields(
-        val fields: List<RegistrationField>,
-        val optionalFields: List<RegistrationField>
-    ) : SignUpUIState()
-    object Loading : SignUpUIState()
-}
+data class SignUpUIState(
+    val fields: List<RegistrationField> = emptyList(),
+    val optionalFields: List<RegistrationField> = emptyList(),
+    val isFacebookAuthEnabled: Boolean = false,
+    val isGoogleAuthEnabled: Boolean = false,
+    val isMicrosoftAuthEnabled: Boolean = false,
+    val isSocialAuthEnabled: Boolean = false,
+    val isLoading: Boolean = false,
+)
