+++ auth/src/main/java/org/openedx/auth/presentation/ui/SocialAuthView.kt
@@ -1,0 +1,121 @@
+package org.openedx.auth.presentation.ui
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import org.openedx.auth.R
+import org.openedx.auth.data.model.AuthType
+import org.openedx.core.ui.OpenEdXButton
+import org.openedx.core.ui.OpenEdXOutlinedButton
+import org.openedx.core.ui.theme.OpenEdXTheme
+import org.openedx.core.ui.theme.appColors
+
+@Composable
+internal fun SocialAuthView(
+    modifier: Modifier = Modifier,
+    isGoogleAuthEnabled: Boolean = true,
+    isFacebookAuthEnabled: Boolean = true,
+    isMicrosoftAuthEnabled: Boolean = true,
+    isSignIn: Boolean = false,
+    onEvent: (AuthType) -> Unit,
+) {
+    Column(modifier = modifier) {
+        if (isGoogleAuthEnabled) {
+            OpenEdXOutlinedButton(
+                modifier = Modifier
+                    .padding(top = 24.dp)
+                    .fillMaxWidth(),
+                backgroundColor = MaterialTheme.appColors.background,
+                borderColor = MaterialTheme.appColors.primary,
+                textColor = Color.Unspecified,
+                onClick = {
+                    onEvent(AuthType.GOOGLE)
+                }
+            ) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_auth_google),
+                        contentDescription = null,
+                        tint = Color.Unspecified,
+                    )
+                    Text(
+                        modifier = Modifier.padding(start = 10.dp),
+                        text = stringResource(id = R.string.auth_google)
+                    )
+                }
+            }
+        }
+        if (isFacebookAuthEnabled) {
+            OpenEdXButton(
+                width = Modifier
+                    .padding(top = 12.dp)
+                    .fillMaxWidth(),
+                text = stringResource(id = R.string.auth_facebook),
+                backgroundColor = MaterialTheme.appColors.authFacebookButtonBackground,
+                onClick = {
+                    onEvent(AuthType.FACEBOOK)
+                }
+            ) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_auth_facebook),
+                        contentDescription = null,
+                        tint = MaterialTheme.appColors.buttonText,
+                    )
+                    Text(
+                        modifier = Modifier.padding(start = 10.dp),
+                        color = MaterialTheme.appColors.buttonText,
+                        text = stringResource(id = R.string.auth_facebook)
+                    )
+                }
+            }
+        }
+        if (isMicrosoftAuthEnabled) {
+            OpenEdXButton(
+                width = Modifier
+                    .padding(top = 12.dp)
+                    .fillMaxWidth(),
+                text = stringResource(id = R.string.auth_microsoft),
+                backgroundColor = MaterialTheme.appColors.authMicrosoftButtonBackground,
+                onClick = {
+                    onEvent(AuthType.MICROSOFT)
+                }
+            ) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_auth_microsoft),
+                        contentDescription = null,
+                        tint = Color.Unspecified,
+                    )
+                    Text(
+                        modifier = Modifier.padding(start = 10.dp),
+                        color = MaterialTheme.appColors.buttonText,
+                        text = stringResource(id = R.string.auth_microsoft)
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Composable
+private fun SocialAuthViewPreview() {
+    OpenEdXTheme {
+        SocialAuthView {}
+    }
+}
